tests: :test_tube: Testing the search bar and sending data correctly
